Renamed folders
pynkTrombone -> dkadish.pynkTrombone

NewPynkTrombone -> pynkTrombone
