+++ pynkTrombone/consts.py
